Official v1.0.9 Alpha Version

+++ Ablockalypse/src/com/github/JamesNorris/Interface/ZAMob.java
@@ -61,11 +61,10 @@ public interface ZAMob {
 
 	/**
 	 * Sets the speed of the entity.
-	 * Default is .03.
 	 * 
 	 * @param speed The speed to set the entity to
 	 */
-	public void setSpeed(double speed);
+	public void setSpeed(float speed);
 
 	/**
 	 * Sets the target of this instance.
